test(Buyer): Buyer 테스트 코드 작성
